backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
bcrypt==5.0.0
billiard==4.2.2
cffi==2.0.0
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
Deprecated==1.3.1
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
kombu==5.5.4
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-dateutil==2.9.0.post0
redis==7.4.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.52.1
tzdata==2026.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.2.14
wrapt==2.1.2
